Update test_pipeline_structure.sh
Updated QC loop to capture R1 (for single-end) and R1/R2 files instead of capturing all .fastq.gz files. This is helpful when we have R3.fastq.gz files containing UMIs which we do not want to use. Updated also echos for printing R1/R2 in the logs out files.